refactor: クラスの名前を「~er」に変更
